% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GwasLocusScores.R
\name{GwasLocusScores}
\alias{GwasLocusScores}
\title{GwasLocusScores}
\description{
assess RNA-binding protein binding sites, their structure and motif matches
}
\section{Fields}{

\describe{
\item{\code{rbp}}{the gene symbol name of the protein}
}}

\examples{
  gls <- GwasLocusScores("rs4575096")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GwasLocusScores$new()}}
\item \href{#method-load.eqtls}{\code{GwasLocusScores$load.eqtls()}}
\item \href{#method-get.tbl.eqtl}{\code{GwasLocusScores$get.tbl.eqtl()}}
\item \href{#method-clone}{\code{GwasLocusScores$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$new(tag.snp, chrom, start.loc, end.loc, tissue.name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag.snp}}{character, an rsid identifying the locus}

\item{\code{chrom}}{character, e.g., "chr1"}

\item{\code{start.loc}}{numeric the start of the genomic region of interest}

\item{\code{end.loc}}{numeric the end of the genomic region of interest}

\item{\code{tissue.name}}{character, e.g. "GTEx.brain_hippocampus"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load.eqtls"></a>}}
\if{latex}{\out{\hypertarget{method-load.eqtls}{}}}
\subsection{Method \code{load.eqtls()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$load.eqtls(pvalue.cutoff)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get.tbl.eqtl"></a>}}
\if{latex}{\out{\hypertarget{method-get.tbl.eqtl}{}}}
\subsection{Method \code{get.tbl.eqtl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$get.tbl.eqtl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GwasLocusScores$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
